pyproject.toml
[project]
name = "tttt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "ddddocr>=1.5.6",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "selenium>=4.35.0",
    "webdriver-manager>=4.0.2",
]
